+++ src/Components/Globe/globe.js
@@ -12,40 +12,39 @@ import {
 
 am4core.useTheme(am4themes_animated);
 
-const GlobeChart = () => {
+const Globe = () => {
   const chartRef = useRef(null);
+  const spinGlobe = useRef(true);
+  const [randomCountries, setRandomCountries] =
+    useRecoilState(randomCountriesState);
+  const [questionCounter, setQuestionCounter] =
+    useRecoilState(questionCounterState);
+  const [quizStarted, setQuizStarted] = useRecoilState(quizStartedState);
+
   useEffect(() => {
-    // am4core.useTheme(am4themes_animated);
+    let chart = am4core.create(chartRef.current, am4maps.MapChart);
 
-    var chart = am4core.create(chartRef.current, am4maps.MapChart);
-
-    // Set map definition
     chart.geodata = am4geodata_worldLow;
-
-    // Set projection
     chart.projection = new am4maps.projections.Orthographic();
     chart.panBehavior = "rotateLongLat";
     chart.deltaLatitude = -20;
     chart.padding(20, 20, 20, 20);
 
-    // Create map polygon series
-    var polygonSeries = chart.series.push(new am4maps.MapPolygonSeries());
-
-    // Make map load polygon (like country names) data from GeoJSON
+    let polygonSeries = chart.series.push(new am4maps.MapPolygonSeries());
     polygonSeries.useGeodata = true;
-    //polygonSeries.include = ["BR", "UA", "MX", "CI"];
 
-    // Configure series
-    var polygonTemplate = polygonSeries.mapPolygons.template;
+    let polygonTemplate = polygonSeries.mapPolygons.template;
     polygonTemplate.tooltipText = "{name}";
-    polygonTemplate.fill = am4core.color("#FF6633");
+    //polygonTemplate.fill = am4core.color("#FF6633");
+    polygonTemplate.fill = am4core.color("#579c49");
+
     polygonTemplate.stroke = am4core.color("#000033");
     polygonTemplate.strokeWidth = 0.5;
-    polygonTemplate.cursorOverStyle = am4core.MouseCursorStyle.pointer;
+    //polygonTemplate.cursorOverStyle = am4core.MouseCursorStyle.pointer;
     polygonTemplate.url = "https://www.datadrum.com/main.php?package={id}";
     polygonTemplate.urlTarget = "_blank";
 
-    var graticuleSeries = chart.series.push(new am4maps.GraticuleSeries());
+    let graticuleSeries = chart.series.push(new am4maps.GraticuleSeries());
     graticuleSeries.mapLines.template.line.stroke = am4core.color("#ffffff");
     graticuleSeries.mapLines.template.line.strokeOpacity = 0.08;
     graticuleSeries.fitExtent = false;
@@ -54,33 +53,78 @@ const GlobeChart = () => {
     chart.backgroundSeries.mapPolygons.template.polygon.fill =
       am4core.color("#ffffff");
 
-    // Create hover state and set alternative fill color
-    var hs = polygonTemplate.states.create("hover");
+    let hs = polygonTemplate.states.create("hover");
     hs.properties.fill = chart.colors.getIndex(0).brighten(-0.5);
 
     let animation;
-    setTimeout(function () {
-      animation = chart.animate(
-        { property: "deltaLongitude", to: 100000 },
-        20000000
-      );
-    }, 3000);
+    if (spinGlobe.current === true) {
+      setTimeout(function () {
+        animation = chart.animate(
+          { property: "deltaLongitude", to: 100000 },
+          20000000
+        );
+      }, 1000);
+    }
 
-    chart.seriesContainer.events.on(
-      "down",
-      function () {
-        //  animation.stop();
-      },
-      []
-    );
+    document.addEventListener("click", function () {
+      animation.stop();
+      spinGlobe.current = false;
+    });
+
+    const desiredColor = am4core.color("#6A5ACD");
+
+    function updateCountry() {
+      if (quizStarted) {
+        const countryName = randomCountries[questionCounter].country;
+
+        polygonSeries.events.on("inited", () => {
+          polygonSeries.mapPolygons.each((polygon) => {
+            if (randomCountries[questionCounter].area < 1500) {
+              //polygon.isHover = true;
+              chart.cursorTooltipEnabled = false; // disable cursor tooltip
+              chart.tooltipContainer.hide(); // hide the cursor tooltip if it is already visible
+              if (!chart.tooltipLabel) {
+                // create tooltip label if it does not exist
+                chart.tooltipLabel = chart.chartContainer.createChild(
+                  am4core.Label
+                );
+                chart.tooltipLabel.fontSize = 12;
+                chart.tooltipLabel.padding(5, 10, 5, 10);
+                chart.tooltipLabel.background.fill = am4core.color("#000000");
+                chart.tooltipLabel.background.fillOpacity = 0.7;
+                chart.tooltipLabel.fill = am4core.color("#ffffff");
+                chart.tooltipLabel.align = "center";
+                chart.tooltipLabel.valign = "middle";
+              }
+              chart.tooltipLabel.text = countryName;
+              chart.tooltipLabel.visible = true;
+              chart.tooltipLabel.x = polygon.visualLongitude;
+              chart.tooltipLabel.y = polygon.visualLatitude - 20; // position label above the country
+            }
+            if (polygon.dataItem.dataContext.name === countryName) {
+              polygon.fill = am4core.color("#6A5ACD");
+            }
+          });
+        });
+        console.log("country name" + countryName);
+        console.log(randomCountries[0].latlng[0]);
+        console.log(randomCountries[0].latlng[1]);
+        chart.deltaLatitude = -1 * randomCountries[questionCounter].latlng[0];
+        chart.deltaLongitude = -1 * randomCountries[questionCounter].latlng[1];
+        //chart.zoomLevel = 1;
+        console.log(chart);
+      }
+    }
+    updateCountry();
+
     return () => {
       if (chart) {
         chart.dispose();
       }
     };
-  }, []);
+  }, [randomCountries, questionCounter]);
 
   return <div ref={chartRef} id="chartdiv" />;
 };
 
-export default GlobeChart;
+export default Globe;
